update(): tie test_opendir & is_permission -> is_valid_path
